Comparators refactor
