+++ spec/comparators/comparator_spec.rb
